requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.7.0
matplotlib
